Media-service/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.14"
fastapi = {extras = ["standart"], version = "^0.128.0"}
uvicorn = {extras = ["standart"], version = "^0.38.0"}
aioboto3 = "^15.5.0"
orjson = "^3.11.7"
pydantic = "^2.12.5"
pydantic-settings = "^2.12.0"
python-multipart = "^0.0.22"
loguru = "^0.7.3"
sqlalchemy = "^2.0.46"
asyncpg = "^0.31.0"
sqlalchemy-utc = "^0.14.0"
alembic = "^1.18.3"
clamd = "^1.0.2"
pyclamd = "^0.4.0"
python-magic-bin = "^0.4.14"
pika = "^1.3.2"
prometheus-fastapi-instrumentator = "^7.1.0"
opentelemetry-api = "^1.39.1"
opentelemetry-sdk = "^1.39.1"
opentelemetry-exporter-otlp = "^1.39.1"
opentelemetry-instrumentation-fastapi = "^0.60b1"
opentelemetry-instrumentation-sqlalchemy = "^0.60b1"
faststream = "^0.6.6"
dishka = "^1.8.0"
aio-pika = "^9.6.1"

[dependency-groups]
dev = [
    "black (>=25.12.0,<26.0.0)"
]
